Add a test with github actions
I wanted to switch from travis ci to github actions, so I added a workflow for testing github actions.

The documentation I referred to is as follows.
https://docs.github.com/ja/enterprise-server@2.22/actions/learn-github-actions/migrating-from-travis-ci-to-github-actions

Since no Ruby 2.0 images were provided for github actions, Ruby 2.0 was removed from the test.
(We consider Ruby 2.0 to be good because it has already reached EOL)

Instead, I added Ruby 3.0 to the test.

The `.travis.yml` will be removed in subsequent commits.